feat(core): extract Arrow adapter crate
