+++ .github/workflows/publish.yml
@@ -22,6 +22,15 @@ jobs:
       - name: Publish geo-polygonize-core to crates.io
         run: cargo publish -p geo-polygonize-core
 
+      - name: Publish geo-polygonize-arrow to crates.io
+        run: |
+          for i in 1 2 3 4 5 6; do
+            cargo publish -p geo-polygonize-arrow && exit 0
+            echo "Publish failed, waiting before retrying..."
+            sleep 30
+          done
+          exit 1
+
       - name: Publish geo-polygonize-wasm to crates.io
         run: |
           for i in 1 2 3 4 5 6; do
