+++ crates/geo-polygonize-arrow/src/lib.rs
@@ -1,0 +1,7 @@
+mod arrow_api;
+pub mod ffi;
+
+#[cfg(feature = "geoparquet")]
+pub mod geoparquet_api;
+
+pub use arrow_api::{polygonize_arrow, PolygonizerOptions};
