+++ crates/geo-polygonize-arrow/tests/test_ffi_errors.rs
@@ -1,7 +1,5 @@
-#![cfg(feature = "arrow")]
-
 use arrow::ffi::{FFI_ArrowArray, FFI_ArrowSchema};
-use geo_polygonize_core::ffi::{polygonize_ffi, PolygonizerOptions};
+use geo_polygonize_arrow::ffi::{polygonize_ffi, PolygonizerOptions};
 
 #[test]
 fn test_ffi_null_pointers() {
