+++ crates/geo-polygonize-core/Cargo.toml
@@ -22,17 +22,13 @@ float_next_after = "1.0"
 smallvec = "1.11"
 pyo3 = { version = "0.29", features = ["extension-module", "abi3-py38"], optional = true }
 numpy = { version = "0.29", optional = true }
-arrow = { version = "57", features = ["ffi"], optional = true }
-geoarrow = { version = "0.7.0", optional = true }
 geo-traits = "0.3.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 humantime-serde = "1.1.1"
 ts-rs = "12.0.1"
-geoparquet = { version = "0.7.0", optional = true }
 flatgeobuf = { version = "6.0.1", optional = true, default-features = false }
 geozero = { version = "0.15.1", optional = true, default-features = false, features = ["with-geo"] }
-parquet = { version = "57.3.0", optional = true }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 multiversion = "0.8.0"
@@ -44,7 +40,6 @@ criterion = "0.5"
 rand = "0.8"
 clap = { version = "4.5", features = ["derive"] }
 geojson = "0.24"
-geoarrow-test = "0.7.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 walkdir = "2.5.0"
@@ -52,8 +47,6 @@ dhat = "0.3.3"
 
 [features]
 default = ["parallel"]
-arrow = ["dep:arrow", "dep:geoarrow"]
-geoparquet = ["arrow", "dep:geoparquet", "dep:parquet"]
 flatgeobuf = ["dep:flatgeobuf", "dep:geozero"]
 parallel = ["dep:rayon"]
 python = ["dep:pyo3", "dep:numpy"]
@@ -77,11 +70,6 @@ harness = false
 [[bench]]
 name = "hole_sort_bench"
 harness = false
-
-[[bench]]
-name = "allocation_bench"
-harness = false
-required-features = ["arrow"]
 
 [[bench]]
 name = "iai_bench"
