+++ release-please-config.json
@@ -28,6 +28,16 @@
         },
         {
           "type": "toml",
+          "path": "crates/geo-polygonize-arrow/Cargo.toml",
+          "jsonpath": "$.package.version"
+        },
+        {
+          "type": "toml",
+          "path": "crates/geo-polygonize-arrow/Cargo.toml",
+          "jsonpath": "$.dependencies.geo-polygonize-core.version"
+        },
+        {
+          "type": "toml",
           "path": "crates/geo-polygonize-wasm/Cargo.toml",
           "jsonpath": "$.package.version"
         },
@@ -35,6 +45,11 @@
           "type": "toml",
           "path": "crates/geo-polygonize-wasm/Cargo.toml",
           "jsonpath": "$.dependencies.geo-polygonize-core.version"
+        },
+        {
+          "type": "toml",
+          "path": "crates/geo-polygonize-wasm/Cargo.toml",
+          "jsonpath": "$.dependencies.geo-polygonize-arrow.version"
         }
       ]
     }
